requirements.txt
django
celery
kombu
billiard
pytz
django-celery
psycopg2
unidecode
unicode_tex
name_tools
-e git://github.com/wetneb/pyoai.git@base_dc#egg=pyoai
redis
django-bulk-update
django-redis-cache
django-cache-machine
python-dateutil
